Edit settings page and tutorial (#37)
* aligned items in footer

* feat: add privacy policy and imprint pages with translations

* feat: enhance dashboard with tutorial steps and command management; update localization for commands

* update dashboard settings UI with slider for guess delay; refactor related styles and logic

* update remove channel functionality with confirmation dialog; localize messages in English and German

* add multilingual support with language flags and localization for German, English, and French

* lint fix

+++ packages/frontend/src/app/pages/dashboard-page/dashboard-page.component.ts
@@ -1,3 +1,4 @@
+import { DecimalPipe } from "@angular/common";
 import { Component, inject, model, type OnInit, signal } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { MatButtonModule } from "@angular/material/button";
@@ -9,8 +10,10 @@ import { MatSnackBar } from "@angular/material/snack-bar";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { Router } from "@angular/router";
 import { TranslatePipe, TranslateService } from "@ngx-translate/core";
+import { CommandComponent } from "../../components/command/command.component";
 import { AuthenticationService } from "../../services/authentication.service";
 import { BackendService } from "../../services/backend.service";
+import { CommandsService } from "../../services/commands.service";
 import { ProfileService } from "../../services/profile.service";
 @Component({
 	selector: "app-dashboard-page",
@@ -23,12 +26,15 @@ import { ProfileService } from "../../services/profile.service";
 		MatSlideToggleModule,
 		FormsModule,
 		MatTooltipModule,
+		CommandComponent,
+		DecimalPipe,
 	],
 	templateUrl: "./dashboard-page.component.html",
 	styleUrl: "./dashboard-page.component.scss",
 })
 export class DashboardPage implements OnInit {
 	protected readonly profileService = inject(ProfileService);
+	protected readonly commandsService = inject(CommandsService);
 	private readonly backendService = inject(BackendService);
 	private readonly authenticationService = inject(AuthenticationService);
 	private readonly router = inject(Router);
@@ -50,10 +56,26 @@ export class DashboardPage implements OnInit {
 			.catch((error) => {
 				const message = this.translate.instant("pages.dashboard.init.error");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 				console.error("Failed to load plugin token:", error);
 				this.pluginToken.set(null);
+				this.authenticationService.removeToken();
+				this.router.navigate(["/landing"]);
+			});
+		this.backendService.channel
+			.me()
+			.then((data) => {
+				this.settingBotActiveWhenOffline.set(data.botActiveWhenOffline);
+				this.settingPublicChannel.set(data.usagePublic);
+				this.settingGuessDelay.set(data.guessDelayTime);
+			})
+			.catch((error) => {
+				const message = this.translate.instant("pages.dashboard.init.error");
+				this.snackbar.open(message, "", {
+					duration: 3000,
+				});
+				console.error("Failed to load channel settings:", error);
 				this.authenticationService.removeToken();
 				this.router.navigate(["/landing"]);
 			});
@@ -65,12 +87,15 @@ export class DashboardPage implements OnInit {
 	}
 
 	protected removeChannel(): void {
+		const confirmation = confirm(this.translate.instant("pages.dashboard.remove.confirmation"));
+		if (!confirmation) return;
+
 		this.backendService.authentication
 			.remove()
 			.then(() => {
 				const message = this.translate.instant("pages.dashboard.remove.success");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 
 				this.authenticationService.removeToken();
@@ -79,7 +104,7 @@ export class DashboardPage implements OnInit {
 			.catch((error) => {
 				const message = this.translate.instant("pages.dashboard.remove.error");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 				console.error("Failed to remove channel:", error);
 			});
@@ -92,14 +117,14 @@ export class DashboardPage implements OnInit {
 			.then((data) => {
 				const message = this.translate.instant("pages.dashboard.token.actions.renew.success");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 				this.pluginToken.set(data.pluginToken);
 			})
 			.catch((error) => {
 				const message = this.translate.instant("pages.dashboard.token.actions.renew.error");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 				console.error("Failed to renew plugin token:", error);
 			});
@@ -114,15 +139,37 @@ export class DashboardPage implements OnInit {
 			.then(() => {
 				const message = this.translate.instant("pages.dashboard.token.actions.copy.success");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 			})
 			.catch((error) => {
 				const message = this.translate.instant("pages.dashboard.token.actions.copy.error");
 				this.snackbar.open(message, "", {
-					duration: 5000,
+					duration: 3000,
 				});
 				console.error("Failed to copy plugin token:", error);
+			});
+	}
+
+	public saveSettings(): void {
+		this.backendService.channel
+			.updateSettings({
+				botActiveWhenOffline: this.settingBotActiveWhenOffline(),
+				usagePublic: this.settingPublicChannel(),
+				guessDelayTime: this.settingGuessDelay(),
+			})
+			.then(() => {
+				const message = this.translate.instant("pages.dashboard.settings.save.success");
+				this.snackbar.open(message, "", {
+					duration: 3000,
+				});
+			})
+			.catch((error) => {
+				const message = this.translate.instant("pages.dashboard.settings.save.error");
+				this.snackbar.open(message, "", {
+					duration: 3000,
+				});
+				console.error("Failed to save settings:", error);
 			});
 	}
 }
